cleaned up some code and changed the view and alerts

+++ src/main/java/com/renaldorasa/sudoku/SudokuGenerator.java
@@ -32,7 +32,7 @@ public class SudokuGenerator
     void fillDiagonal()
     {
 
-        for (int i = 0; i<N; i=i+SRN)
+        for (int i = 0; i < N; i = i + SRN)
             fillBox(i, i);
     }
 
@@ -40,7 +40,7 @@ public class SudokuGenerator
     {
         for (int i = 0; i<SRN; i++)
             for (int j = 0; j<SRN; j++)
-                if (mat[rowStart+i][colStart+j]==num)
+                if (mat[rowStart + i][colStart + j] == num)
                     return false;
 
         return true;
@@ -49,9 +49,9 @@ public class SudokuGenerator
     void fillBox(int row,int col)
     {
         int num;
-        for (int i=0; i<SRN; i++)
+        for (int i = 0; i < SRN; i++)
         {
-            for (int j=0; j<SRN; j++)
+            for (int j = 0; j < SRN; j++)
             {
                 do
                 {
@@ -59,7 +59,7 @@ public class SudokuGenerator
                 }
                 while (!unUsedInBox(row, col, num));
 
-                mat[row+i][col+j] = num;
+                mat[row + i][col + j] = num;
             }
         }
     }
@@ -69,16 +69,16 @@ public class SudokuGenerator
         return (int) Math.floor((Math.random()*num+1));
     }
 
-    boolean CheckIfSafe(int i,int j,int num)
+    boolean CheckIfSafe(int i, int j, int num)
     {
         return (unUsedInRow(i, num) &&
                 unUsedInCol(j, num) &&
-                unUsedInBox(i-i%SRN, j-j%SRN, num));
+                unUsedInBox(i - i % SRN, j - j % SRN, num));
     }
 
     boolean unUsedInRow(int i,int num)
     {
-        for (int j = 0; j<N; j++)
+        for (int j = 0; j < N; j++)
             if (mat[i][j] == num)
                 return false;
         return true;
@@ -87,7 +87,7 @@ public class SudokuGenerator
     // check in the row for existence
     boolean unUsedInCol(int j,int num)
     {
-        for (int i = 0; i<N; i++)
+        for (int i = 0; i < N; i++)
             if (mat[i][j] == num)
                 return false;
         return true;
@@ -95,12 +95,12 @@ public class SudokuGenerator
 
     boolean fillRemaining(int i, int j)
     {
-        if (j>=N && i<N-1)
+        if (j >= N && i < N-1)
         {
             i = i + 1;
             j = 0;
         }
-        if (i>=N && j>=N)
+        if (i >= N && j >= N)
             return true;
 
         if (i < SRN)
@@ -108,23 +108,23 @@ public class SudokuGenerator
             if (j < SRN)
                 j = SRN;
         }
-        else if (i < N-SRN)
+        else if (i < N - SRN)
         {
-            if (j==(int)(i/SRN)*SRN)
+            if (j == (i / SRN) * SRN)
                 j = j + SRN;
         }
         else
         {
-            if (j == N-SRN)
+            if (j == N - SRN)
             {
                 i = i + 1;
                 j = 0;
-                if (i>=N)
+                if (i >= N)
                     return true;
             }
         }
 
-        for (int num = 1; num<=N; num++)
+        for (int num = 1; num <= N; num++)
         {
             if (CheckIfSafe(i, j, num))
             {
@@ -143,10 +143,10 @@ public class SudokuGenerator
         int count = K;
         while (count != 0)
         {
-            int cellId = randomGenerator(N*N - 1);
+            int cellId = randomGenerator(N * N - 1);
 
-            int i = (cellId/N);
-            int j = cellId%9;
+            int i = (cellId / N);
+            int j = cellId % 9;
             if (j != 0)
                 j = j - 1;
 
@@ -163,9 +163,9 @@ public class SudokuGenerator
     {
         StringBuilder board = new StringBuilder();
 
-        for (int i = 0; i<N; i++)
+        for (int i = 0; i < N; i++)
         {
-            for (int j = 0; j<N; j++)
+            for (int j = 0; j < N; j++)
                 board.append(mat[i][j]);
         }
         return board.toString();
